+++ Headers/cuda_ops.hpp
@@ -4,10 +4,10 @@
 #include "globals.hpp"
 
 namespace cuda {
-    ::std::vector<double> matmul(const ::std::vector<double>& matrix_A, const ::std::vector<double>& matrix_B, int K, int X, int Y);
-    ::std::vector<double> binary_op(const char op, const ::std::vector<double>& matrix_A, const ::std::vector<double>& matrix_B, int size);
-    void softmax(const ::std::vector<double>& input, double* output, int N, int C);
+    ::std::vector<float> matmul(const ::std::vector<float>& matrix_A, const ::std::vector<float>& matrix_B, int K, int X, int Y);
+    ::std::vector<float> binary_op(const char op, const ::std::vector<float>& matrix_A, const ::std::vector<float>& matrix_B, int size);
+    void softmax(const ::std::vector<float>& input, float* output, int N, int C);
 
-    double reduction(const ReductionOp op, const ::std::span<const double>& input);
-    double full_reduction(const ReductionOp op, const ::std::span<const double>& input);
+    float reduction(const ReductionOp op, const ::std::span<const float>& input);
+    float full_reduction(const ReductionOp op, const ::std::span<const float>& input);
 }
